test: move cmap setup and cleanup to runner file
